Updated to use an event and command handler
Updated to use an event and command handler rather than hardcoding all of
the events and commands to watch for in the bot.js file. Now all that
needs to be done is add a new file named the event or command's name with
the code cooresponding with it in the related folder and update the
eventLoader.js file for events.

Also added an echo command and the bones of a timer command and logging of
disconnect and reconnect events of the bot from the server.